refactor: PR review - theme 분리 및 spacing 적용
