+++ src/App.tsx
@@ -1,6 +1,6 @@
 import GlobalStyle from "@/styles/GlobalStyle";
 import { ThemeProvider } from "@emotion/react";
-import theme from "@/styles/theme";
+import theme from "@/styles/theme/theme";
 import Gift from "@/pages/Gift";
 
 function App() {
