+++ src/assets/ranking_item_mock.ts
@@ -2,8 +2,7 @@ export const ranking_item_mock = [
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -12,15 +11,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -29,15 +26,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -46,15 +41,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -63,15 +56,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -80,15 +71,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -97,15 +86,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -114,15 +101,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -131,15 +116,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -148,15 +131,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -165,15 +146,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -182,15 +161,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -199,15 +176,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -216,15 +191,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -233,15 +206,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -250,15 +221,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -267,15 +236,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -284,15 +251,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -301,15 +266,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -318,15 +281,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -335,15 +296,13 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
   {
     id: 123,
     name: "BBQ 양념치킨+크림치즈볼+콜라1.25L",
-    imageURL:
-      "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
+    imageURL: "https://st.kakaocdn.net/product/gift/product/20231030175450_53e90ee9708f45ffa45b3f7b4bc01c7c.jpg",
     price: {
       basicPrice: 29000,
       discountRate: 0,
@@ -352,8 +311,7 @@ export const ranking_item_mock = [
     brandInfo: {
       id: 2088,
       name: "BBQ",
-      imageURL:
-        "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
+      imageURL: "https://st.kakaocdn.net/product/gift/gift_brand/20220216170226_38ba26d8eedf450683200d6730757204.png",
     },
   },
 ];
