+++ src/components/DividerDiv.tsx
@@ -7,7 +7,7 @@ const DividerDiv = () => {
 const Divider = styled.div`
   width: 100%;
   height: 30px;
-  background-color: ${({ theme }) => theme.background_color.default};
+  background-color: ${({ theme }) => theme.color.backgroundColor.default};
 `;
 
 export default DividerDiv;
