+++ src/components/RankingList.tsx
@@ -2,11 +2,13 @@ import { ranking_item_mock } from "@/assets/ranking_item_mock";
 import styled from "@emotion/styled";
 import DividerDiv from "@/components/DividerDiv";
 import { useState } from "react";
-import type { ThemeType } from "@/styles/theme";
+import type { ThemeType } from "@/styles/theme/theme";
 import { useTheme } from "@emotion/react";
 
+const rankingListViewCount = 6;
+
 const RankingList = () => {
-  const [viewCount, setViewCount] = useState(6);
+  const [viewCount, setViewCount] = useState(rankingListViewCount);
   const theme = useTheme();
   return (
     <Container>
@@ -32,12 +34,10 @@ const RankingList = () => {
       <ItemContent>
         <MoreBtn
           onClick={() => {
-            viewCount == 6
-              ? setViewCount(ranking_item_mock.length)
-              : setViewCount(6);
+            viewCount == rankingListViewCount ? setViewCount(ranking_item_mock.length) : setViewCount(6);
           }}
         >
-          {viewCount == 6 ? "더보기" : "접기"}
+          {viewCount == rankingListViewCount ? "더보기" : "접기"}
         </MoreBtn>
       </ItemContent>
       <DividerDiv />
@@ -52,7 +52,7 @@ const Content = styled.div`
   width: 100%;
   display: grid;
   grid-template-columns: repeat(3, 1fr);
-  gap: 24px 8px;
+  gap: ${({ theme }) => theme.spacing.spacing6} ${({ theme }) => theme.spacing.spacing2};
 `;
 const Item = styled.div`
   width: 100%;
@@ -73,12 +73,9 @@ const ItemRank = styled.span<RankingAndTheme>`
   justify-content: center;
   top: 0.25rem;
   left: 0.25rem;
-  font: ${({ theme }) => theme.typography.label2Bold};
-  color: ${({ theme }) => theme.colors.gray00};
-  background-color: ${(props) =>
-    props.ranking <= 3
-      ? props.theme.colors.red500
-      : props.theme.colors.gray700};
+  font: ${({ theme }) => theme.typography.label1Bold};
+  color: ${({ theme }) => theme.color.gray00};
+  background-color: ${(props) => (props.ranking <= 3 ? props.theme.color.red500 : props.theme.color.gray700)};
 `;
 const ItemContent = styled.div`
   width: 100%;
@@ -93,32 +90,32 @@ const ItemContentImg = styled.img`
   object-position: center center;
   border-radius: 4px;
   overflow: hidden;
-  margin-bottom: 10px;
+  margin-bottom: ${({ theme }) => theme.spacing.spacing3};
 `;
 const ItemContentBrand = styled.p`
   font: ${({ theme }) => theme.typography.subtitle2Regular};
-  color: ${({ theme }) => theme.text_color.sub};
+  color: ${({ theme }) => theme.color.textColor.sub};
   margin-right: auto;
 `;
 const ItemContentTitle = styled.h6`
   font: ${({ theme }) => theme.typography.subtitle2Bold};
-  color: ${({ theme }) => theme.text_color.default};
+  color: ${({ theme }) => theme.color.textColor.default};
   margin-right: auto;
-  margin-bottom: 5px;
+  margin-bottom: ${({ theme }) => theme.spacing.spacing1};
 `;
 const ItemContentPrice = styled.p`
   font: ${({ theme }) => theme.typography.title2Bold};
-  color: ${({ theme }) => theme.text_color.default};
+  color: ${({ theme }) => theme.color.textColor.default};
   margin-right: auto;
   word-break: break-word;
 `;
 const MoreBtn = styled.button`
   max-width: 30rem;
   width: 100%;
-  padding: 12px;
+  padding: ${({ theme }) => theme.spacing.spacing3};
   border-radius: 4px;
-  border: 1px solid ${({ theme }) => theme.colors.gray400};
-  background-color: ${({ theme }) => theme.colors.gray00};
+  border: 1px solid ${({ theme }) => theme.color.gray400};
+  background-color: ${({ theme }) => theme.color.gray00};
   cursor: pointer;
 `;
