+++ src/styles/theme/emotion.d.ts
@@ -1,5 +1,5 @@
 import "@emotion/react";
-import type { ThemeType } from "@/styles/theme";
+import type { ThemeType } from "@/styles/theme/theme";
 
 declare module "@emotion/react" {
   export interface Theme extends ThemeType {}
