+++ src/styles/theme/spacing.ts
@@ -1,0 +1,21 @@
+const spacing = {
+  spacing0: "0px", //간격 없음
+  spacing1: "4px", //최소 간격
+  spacing2: "8px", //작은 간격
+  spacing3: "12px", //작은-중간 간격
+  spacing4: "16px", //기본 간격
+  spacing5: "20px", //중간 간격
+  spacing6: "24px", //중간-큰 간격
+  spacing7: "28px", //큰 간격
+  spacing8: "32px", //매우 큰 간격
+  spacing9: "36px", //초대형 간격
+  spacing10: "40px", //40px 간격
+  spacing11: "44px", //44px 간격
+  spacing12: "48px", //48px 간격
+  spacing13: "52px", //52px 간격
+  spacing14: "56px", //56px 간격
+  spacing15: "60px", //60px 간격
+  spacing16: "64px", //최대 간격
+};
+
+export default spacing;
