+++ src/styles/theme/theme.ts
@@ -1,0 +1,12 @@
+import color from "@/styles/theme/color";
+import typography from "@/styles/theme/typography";
+import spacing from "@/styles/theme/spacing";
+
+const theme = {
+  color,
+  typography,
+  spacing,
+};
+
+export type ThemeType = typeof theme;
+export default theme;
