+++ src/styles/theme/typography.ts
@@ -1,0 +1,27 @@
+const typography = {
+  title1Bold: '700 1.25rem/1.6875rem "Pretendard", sans-serif',
+  title1Regular: '400 1.25rem/1.6875rem "Pretendard", sans-serif',
+
+  title2Bold: '700 1rem/1.5rem "Pretendard", sans-serif',
+  title2Regular: '400 1rem/1.5rem "Pretendard", sans-serif',
+
+  subtitle1Bold: '700 1rem/1.375rem "Pretendard", sans-serif',
+  subtitle1Regular: '400 1rem/1.5rem "Pretendard", sans-serif',
+
+  subtitle2Bold: '700 0.875rem/1.1875rem "Pretendard", sans-serif',
+  subtitle2Regular: '400 0.875rem/1.1875rem "Pretendard", sans-serif',
+
+  body1Bold: '700 1rem/1.375rem "Pretendard", sans-serif',
+  body1Regular: '400 1rem/1.375rem "Pretendard", sans-serif',
+
+  body2Bold: '700 0.875rem/1.1875rem "Pretendard", sans-serif',
+  body2Regular: '400 0.875rem/1.1875rem "Pretendard", sans-serif',
+
+  label1Bold: '700 0.875rem/1.1875rem "Pretendard", sans-serif',
+  label1Regular: '400 0.875rem/1.1875rem "Pretendard", sans-serif',
+
+  label2Bold: '700 0.75rem/1rem "Pretendard", sans-serif',
+  label2Regular: '400 0.75rem/1rem "Pretendard", sans-serif',
+};
+
+export default typography;
